doc: setup instruction about Glide
For gophers who are not used to Glide, it's unclear what we should do
with the repos once fetched and all dependencies has been installed.

Make it clear that it still can be used in a classic go workspace and
users do not have to use Glide in their own projects.