horizon/cmd: fix configuring logging level (#979)
This issue is caused by reverting a PR on the release branch for quickly unblocking, not caused by the logrus library update. That temporary PR was brought back to master after we published the release, and the PR reverting the revert didn't clean it up everything properly. Thus we ended up in this state.

This PR uses Logger.SetLevel to properly set the level of the logger in an Entry.