keystore/cmd: add tls config to keystore (#1243)
This PR adds tls config to the keystore server. It copies the regen script and the localhost config from Horizon. Of course, it updates the config file so that it points to the right port.

After istalling keystored, one can do keystored -tls-cert=path_to_cert -tls-key=path_to_key serve to start keystored. I will add a README as the next step.