exp/keystore: implement storeKeys endpoint (#1326)
This PR adds the very first keystore endpoint for storing keys. Users are expected to encrypt their keys on the client side and send the keys_blob together with the encrypter_name and salt. We expect the client to encode the blob as base64 raw url string. Each user can only have exact one row in our database, meaning that at this moment they cannot encrypt the same key with a different encrypter.

This PR introduces a recoverHandler to work with the jsonHandler because jsonHandler can panic if the developer doesn't follow the allowed function prototype in our support/render/hal package. There are more handlers to be added in the wrapMiddleware like authHandler where we forward the request to the configured auth endpoint and timeoutHandler, traceHandler, etc.

This PR implements two tests for storing keys, one inside the service, one at the http level.