support/render/hal: relax the return value constraint (#1373)
This PR relaxes the two-return-value constraint in the function type check. We will return a default response if there's no return value. If there's one return value, we will return something based on whether that single return value is an error or not.

The motivation behind this improvement is that in keystore we have a HTTP DELETE endpoint, where we only return a default response if it's successful.