exp/ingest: Verification of buckets in history archive (#1502)
This commit provides verification of buckets in history archive
exploiting the fact that hash of buckets for HAS is stored in
`LedgerHeader`.

History archives cannot be fully trusted if they are maintained by a
third party or are not exposed using a safe protocol. The goal of this
PR is to add verification of buckets to ensure data is safe and has not
been changed. The verification works like this:

0. The prerequisite is a trusted Stellar-Core node that will provide an
authentic ledger header (or at least header of the checkpoint ledger
obtained from a trusted source).
1. We get ledger header from Stellar-Core backend for the HAS we want to
apply.
2. We calculate bucket list hash using data in HAS and compare it with
the authentic ledger header from 1.
3. For each bucket file we check if the hash of it's contents matches
the hash in the file name.

This way any change in HAS or in any of buckets found in HAS will
either: change the bucket list hash (will not be equal to the one in
stellar-core) or will change the bucket file hash (and file name) so
this will affect the list of buckets in HAS (and as a result change the
bucket list hash).