exp/ingest: TempSet in SingleLedgerStateReader (#1539)
This commit adds a new `TempSet` implementations (`MemoryTempSet` and
`PostgresTempSet`) for holding temporary objects during processing of
history buckets.

The existing in-memory method for storing temporary object (keys seen in
previous buckets) when processing history buckets is fast but requires
~4GB of RAM to process public network state. This requirement makes it
impossible to ingest the state on machines with less memory.

To solve this we add an alternative `PostgresTempSet` that dumps the
keys to a temporary Postgres table when number of keys in memory exceed
the limit. It still operates on in-memory cache but the requirements for
it are much smaller than in `MemoryTempSet`. The objects are loaded and
saved in batches what reduces round trip time in DB connection. To
achieve this, `BucketEntry`ies read from history buckets are buffered
(instead of reading one by one) and the keys are preloaded from a store.
Additionally, to save memory `MarshalBinaryCompress` method has been
implemented on some types in `xdr` package to make the marshalled
`LedgerKey` smaller. The compression exploits the fact that XDR encoding
inserts `0` bytes for padding which are not needed in this use case.

This commit also includes some refactoring, ex. `MemoryStateReader` was
renamed to `SingleLedgerStateReader` which better explains it.

Finally, some helper scripts that compare results produced by new code
with the state in stellar-core instance have been added.