services/horizon: Distributed ingestion (#1566)
This PR changes new ingestion system in Horizon to allow distributed
ingestion. This means that ingestion will be running on all Horizon
nodes but only one, random node will be a leader that updates a database
(at the same time all of the nodes will update in-memory order book
graph).

In #1519 we added in-memory order book graph that allows faster path
finding. The issue we noticed is that the graph is updated by ingesting
pipeline which runs on a single Horizon instance only. To solve this we
decided to run ingestion on all nodes. This will also improve
availability of Horizon: if one node goes down other nodes can continue
ingestion.

We need to support cartesian product of the following run scenarios
cases:
- Init from empty state (1a) and resuming from existing state (1b).
- Ingestion system version has been upgraded (2a) or not (2b).
- Current node is leading ingestion (3a) or not (3b).

We always clear state when ingestion system is upgraded so 2a and 2b are
included in 1a.

We ensure that only one instance is a leader because in each round
instances try to acquire a lock on `LastLedgerExpIngest` value in key
value store and only one instance will be able to acquire it. This
happens in both initial processing and ledger processing. So this solves
3a and 3b in both 1a and 1b.
 
Finally, 1a and 1b are tricky because we need to keep the latest version
of order book graph in memory of each Horizon instance. To solve this:
* For state init:
  * If instance is a leader, we update the order book graph by running
    state pipeline normally.
  * If instance is NOT a leader, we build a graph from offers present in
    a database. We completely omit state pipeline in this case.
* For resuming:
  * If instances is a leader, it runs full ledger pipeline, including
    updating a database.
  * If instances is a NOT leader, it runs ledger pipeline without
    updating a database so order book graph is updated but database is
    not overwritten.