services/horizon: Fix upgrade when database not empty (#1574)
This commit fixes a scenario when a state is cleared because of the
ingestion system upgrade. In the pipelines post-processing hook we have
an extra check to ensure that the ledger sequence of the last locally
processed ledger is exactly +1 the the ledger sequence of the last
processed ledger in a database. This fails for init state if there are
some ledgers ingested because LastLedgerExpIngest was not reset during
clearing state.