all: drop support for go 1.10 (#1608)
* all: drop support for go 1.10

Drop support for Go 1.10, which makes the official supported versions:
 - Go 1.11
 - Go 1.12

The officially supported versions of Go are only the last two releases.
At this time the last release is Go 1.12, and so only that version and
the version prior are supported. Supported versions of Go get bug fixes
and security updates, but versions prior do not. The code in this
repository is used in critical systems and we should encourage its use
only in versions of Go that are receiving security updates.

Additionally we plan to move to using Go Modules instead of Go Dep for
dependency management. Go Modules are only supported in Go 1.11 onwards.

* all: update changelogs

* all: pin Dockerfile to latest release of go1.12 instead of a specific patch version

* all: temporarily re-add the jobs so they will no-op and github will let me merge

* all: temporarily re-add the jobs being removed

* all: temporarily re-add the jobs being removed

* all: remove the jobs from the workflow

* all: remove the jobs again