keystore: implement keystore based on the new spec (#1618)
This PR changes the keystore implementation to reflect on the latest spec.

Please note that the sent keysBlob will not necessary equal to the one returned because we store the decoded blob as jsonb type in the database. Storing data in such type will benefit us for having more flexibility for future features. For example, we might want to be able to query a specific keys by its ID.

This PRs also updates rubenv/sql-migrate package to resolve a migration order issue. The new version of the package finds an error in the tickerdb migration. The fix is included in this PR.